refactor(move): Extract LoaderModal into components
The purpose of this is to reuse it in other components that are not linked to the move functionality
